Remove unnecessary methods/classes & bug fixes

+++ twitter-service/src/main/java/com/bot/discord/server/DiscordServerController.java
@@ -11,8 +11,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+@RestController
 public class DiscordServerController {
 	
 	@Autowired
@@ -33,14 +35,20 @@ public class DiscordServerController {
 		return server.get();
 	}
 	
-	@DeleteMapping("/twitter/listeners/{id}")
+	@DeleteMapping("/discord/servers/{id}")
 	public void deleteUser(@PathVariable long id) {
+		Optional<DiscordServer> server = serverRepository.findById(id);
+		
+		if(!server.isPresent())
+			throw new DiscordServerNotFoundException("id-" + id);
+		
 		serverRepository.deleteById(id);
 	}
 	
-	@PostMapping("/twitter/listeners")
-	public ResponseEntity<Object> createListener(@RequestBody DiscordServer server) {
-		DiscordServer savedDiscord = serverRepository.save(server);
+	@PostMapping("/discord/servers")
+	public ResponseEntity<Object> createListener(@RequestBody DiscordServerDTO server) {
+		DiscordServer discordServer = new DiscordServer(server);
+		DiscordServer savedDiscord = serverRepository.save(discordServer);
 		
 		URI location = ServletUriComponentsBuilder
 				.fromCurrentRequest()
